requirements.txt
TurboGears2
setuptools
google-cloud
oauth2client==2.2.0